FIX Increase the scope we run TravisCI with
Currently we're planning to release this module in a recipe that requires
the core at versions 4.1 - this module currently tests against 4.0 and is
all green (pass). The 4.1 upgrade though sees a number of those tests
suddenly turn red (fail) and we need visibility on this before we get to
stabilising against the new core versions.